video/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0

# Database and ORM
sqlmodel==0.0.14
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography>=41.0.0,<46.0.0

# HTTP client and requests
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1
tenacity

# Data processing and validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
vaderSentiment

# Video and media processing
moviepy==1.0.3
imageio-ffmpeg==0.4.9
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.4

# Machine Learning and AI
pandas==2.1.4
scikit-learn==1.3.2
torch==2.1.1
transformers==4.36.0

# Monitoring and observability
sentry-sdk[fastapi]==1.38.0
posthog==3.0.2
json-log-formatter

# Task queue and background processing
celery==5.3.4
redis==5.0.1  # Also used for rate limiting

# Testing
pytest==7.4.3
pytest-httpx
pytest-asyncio==0.21.1
pytest-cov==4.0.0
coverage==7.3.2

# Load Testing
locust==2.17.0
aiohttp==3.9.1

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
python-magic
safety
bandit

# Streamlit for dashboard
streamlit==1.28.2
plotly==5.17.0
altair==5.2.0

# System monitoring
psutil==5.9.6
prometheus-fastapi-instrumentator==6.1.0

# Cloud storage (S3/MinIO)
boto3==1.34.0
botocore==1.34.0

# Supabase client
supabase==2.3.4

# Additional utilities
python-dateutil==2.8.2
pytz==2023.3